+++ lib/astunparse/__init__.py
@@ -1,13 +1,20 @@
 # coding: utf-8
 from __future__ import absolute_import
 from six.moves import cStringIO
-from ._unparse import Unparser
+from .unparser import Unparser
+from .printer import Printer
 
 
-__version__ = '1.1.0'
+__version__ = '1.2.0'
 
 
 def unparse(tree):
     v = cStringIO()
     Unparser(tree, file=v)
+    return v.getvalue()
+
+
+def dump(tree):
+    v = cStringIO()
+    Printer(file=v).visit(tree)
     return v.getvalue()